Updated routed and controller for search EndPoint

+++ app/controllers/freelancerBannerProjects.controller.js
@@ -3,6 +3,7 @@ const FreelancerBannerProject = db.freelancerBannerProject;
 const Registration = db.registration;
 const path = require("path");
 const fs = require("fs");
+const { error } = require("console");
 require('dotenv').config();
 const USERS_BANNER_PATH = path.join(process.env.USERS_BANNER_PATH);
 
@@ -17,7 +18,9 @@ module.exports.createUsersBanner = async function (req, res) {
     let registrationId = req.params.registrationId;
     const registrationRecord = await Registration.findByPk(registrationId);
     if (!registrationRecord || registrationRecord.registration_type !== 3) {
-        fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', req.files['images'][0].filename));
+        if(req.files['images']){
+            fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', req.files['images'][0].filename));
+        }
         return res.status(404).json({ error: 'Registration not exist or Registration is not freelancer type' });
     }
     const { banner_name } = req.body;
@@ -25,7 +28,11 @@ module.exports.createUsersBanner = async function (req, res) {
         fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', req.files['images'][0].filename));
         return res.status(400).json({ error: 'Please Provide Banner Name' });
     }
-    const bannerImages = req.files['images'];
+    let bannerImages;
+    if(req.files['images']){
+        bannerImages = req.files['images'];
+    }
+     
     if (bannerImages && bannerImages.length > 0) {
         const userBanner = await FreelancerBannerProject.create({
             banner_name: banner_name,
@@ -51,23 +58,59 @@ module.exports.createUsersBanner = async function (req, res) {
 module.exports.updateUsersBanner = async function (req, res) {
     let userBannerId = req.params.userBannerId;
     let userBanner = await FreelancerBannerProject.findByPk(userBannerId);
-    const { banner_name } = req.body;
-    if (!banner_name) {
-        fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', req.files['images'][0].filename));
-        return res.status(400).json({ error: 'Please Provide Banner Name' });
+    if(!userBanner){
+        return res.status(404).json({ error :'No record found with provided userBannerId'})
     }
-    const bannerImages = req.files['images'];
+    const { banner_name,registrationId } = req.body;
+
+    const registrationRecord = await Registration.findByPk(registrationId);
+    if (!registrationRecord || registrationRecord.registration_type !== 3) {
+        if(req.files['images']){
+            fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', req.files['images'][0].filename));
+        }
+        return res.status(404).json({ error: 'Registration not exist or Registration is not freelancer type' });
+    }
+    
+    let bannerImages;
+    if(req.files['images']){
+        bannerImages = req.files['images'];
+    }
     if (bannerImages && bannerImages.length > 0) {
         if (userBanner.banner_image) {
             fs.unlinkSync(path.join(__dirname, '../..', USERS_BANNER_PATH, '/', userBanner.banner_image));
         }
-        userBanner.banner_image = bannerImages[0].filename;
-        userBanner.banner_name = banner_name
-        await userBanner.save();
-        return res.status(200).json({ message: "UserBanner Updated Successfully", data: userBanner });
+        const [row,bannerRecord] = await FreelancerBannerProject.update({
+            banner_image : bannerImages[0].filename,
+            banner_name : banner_name,
+            registrationId: registrationId
+        },{
+            where : {
+                id : userBannerId
+            },
+            returning: true
+        });
+        if(row>0){
+            return res.status(200).json({ message: "UserBanner Updated Successfully", data: bannerRecord[0] });
+        }else{
+            return res.status(400).json({error: 'Record not updated'})
+        }
+        
     }
     else{
-        return res.status(400).json({ error : 'Banner File not provided'})
+        const [row,bannerRecord] = await FreelancerBannerProject.update({
+            banner_name : banner_name,
+            registrationId: registrationId
+        },{
+            where : {
+                id : userBannerId
+            },
+            returning : true
+        });
+        if(row>0){
+            return res.status(200).json({ message: "UserBanner Updated Successfully", data: bannerRecord[0] });
+        }else{
+            return res.status(400).json({error: 'Record not updated'})
+        }
     }
     
 }
@@ -106,7 +149,7 @@ module.exports.getUsersBannerByRegistrationId = async function (req, res) {
     if (bannerUserRecord) {
         return res.status(200).json({ message: 'UsersBanner Record with Registration Id', data: bannerUserRecord });
     } else {
-        return res.status(400).json({ error: 'No UserBanner Record with id Present' });
+        return res.status(400).json({ error: 'No UserBanner Record with Registration id Present' });
     }
 }
 
@@ -141,3 +184,58 @@ module.exports.deleteUsersBanner = async function (req, res) {
         return res.status(401).json({ error: 'No UserBanner Deleted' });
     }
 }
+
+/**
+ * Search User Banner details by fieldName and  fieldValue from the database.
+ *
+ * @param {Object} req - Express request object.
+ * @param {Object} res - Express response object.
+ * @returns {Promise<void>} - Promise representing the completion of the retrieval operation.
+ */
+
+module.exports.search = async function (req, res) {
+    try {
+        const { fieldName, fieldValue } = req.params
+        if (!FreelancerBannerProject.rawAttributes[fieldName]) {
+            return res.status(400).json({ error: 'Invalid field name' });
+        }
+        const records = await FreelancerBannerProject.findAll({
+            where: {
+                [fieldName]: fieldValue,
+            },
+        });
+        if (records.length > 0) {
+            return res.status(200).json({ message: 'Fetched Records', data: records })
+        } else {
+            return res.status(404).json({ error: 'No record found' })
+        }
+
+    } catch (err) {
+        if (err instanceof Sequelize.Error) {
+            return res.status(400).json({ error: err.message })
+        }
+        return res.status(500).json({ error: 'Internal Server Error' })
+    }
+}
+
+
+/**
+ * Retrieve all Banner  details from the database.
+ *
+ * @param {Object} req - Express request object.
+ * @param {Object} res - Express response object.
+ * @returns {Promise<void>} - Promise representing the completion of the retrieval operation.
+ */
+
+module.exports.getAll = async function (req, res) {
+    try {
+        const records = await FreelancerBannerProject.findAll({});
+        if (records.length > 0) {
+            return res.status(200).json({ message: "Details fetched successfully", data: records });
+        } else {
+            return res.status(404).json({ error: "details not found" });
+        }
+    } catch (err) {
+        return res.status(500).json({ error: "Internal Server Error" });
+    }
+};
